+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Include migration files in package
+- Add tests for udata-triggered checks [#49](https://github.com/etalab/udata-hydra/pull/49)
 
 ## 1.0.1 (2023-01-04)